Restructured the protocol
