+++ protocol/judge/commands/__init__.py
@@ -1,0 +1,9 @@
+from enum import Enum
+
+from .check_command import CheckCommand
+from .start_command import StartCommand
+
+
+class Commands(Enum):
+    START = StartCommand()
+    CHECK = CheckCommand()
