+++ protocol/judge/commands/start_command.py
@@ -1,0 +1,8 @@
+from .command import Command
+
+
+class StartCommand(Command):
+    @staticmethod
+    def response(message: dict):
+        pass
+
