+++ protocol/runner/commands/__init__.py
@@ -1,0 +1,7 @@
+from enum import Enum
+
+from .check_command import CheckCommand
+
+
+class Commands(Enum):
+    CHECK = CheckCommand()
